Move terminal widget logic to the separed extension.
