+++ .travis.yml
@@ -59,6 +59,7 @@ cache:
   - packages/userstorage/node_modules
   - packages/variable-resolver/node_modules
   - packages/workspace/node_modules
+  - packages/xterm-js/node_modules
 # end_cache_directories
   - packages/java-debug/download
   - packages/debug-nodejs/download
