+++ examples/browser/package.json
@@ -52,7 +52,8 @@
     "@theia/typescript": "^0.3.19",
     "@theia/userstorage": "^0.3.19",
     "@theia/variable-resolver": "^0.3.19",
-    "@theia/workspace": "^0.3.19"
+    "@theia/workspace": "^0.3.19",
+    "@theia/xterm.js": "^0.3.19"
   },
   "scripts": {
     "prepare": "yarn run clean && yarn build",
