+++ examples/electron/package.json
@@ -55,7 +55,8 @@
     "@theia/typescript": "^0.3.19",
     "@theia/userstorage": "^0.3.19",
     "@theia/variable-resolver": "^0.3.19",
-    "@theia/workspace": "^0.3.19"
+    "@theia/workspace": "^0.3.19",
+    "@theia/xterm.js": "^0.3.19"
   },
   "scripts": {
     "prepare": "yarn run clean && yarn build",
