+++ packages/task/src/browser/task-frontend-contribution.ts
@@ -22,7 +22,7 @@ import { FrontendApplication, FrontendApplicationContribution, QuickOpenContribu
 import { WidgetManager } from '@theia/core/lib/browser/widget-manager';
 import { TaskContribution, TaskResolverRegistry, TaskProviderRegistry } from './task-contribution';
 import { TaskService } from './task-service';
-import { TerminalMenus } from '@theia/terminal/lib/browser/terminal-frontend-contribution';
+import { TerminalMenus } from '@theia/xterm.js/lib/browser/terminal-frontend-contribution';
 
 export namespace TaskCommands {
     const TASK_CATEGORY = 'Task';
