+++ packages/terminal/src/browser/terminal-service.ts
