+++ packages/terminal/src/browser/terminal-widget.ts
