+++ packages/xterm-js/src/browser/index.ts
