+++ packages/xterm-js/src/browser/terminal-frontend-contribution.ts
@@ -34,8 +34,8 @@ import { TabBarToolbarContribution, TabBarToolbarRegistry } from '@theia/core/li
 import { WidgetManager } from '@theia/core/lib/browser';
 import { TERMINAL_WIDGET_FACTORY_ID, TerminalWidgetFactoryOptions } from './terminal-widget-impl';
 import { TerminalKeybindingContexts } from './terminal-keybinding-contexts';
-import { TerminalService } from './base/terminal-service';
-import { TerminalWidgetOptions, TerminalWidget } from './base/terminal-widget';
+import { TerminalService } from '@theia/terminal/lib/browser/terminal-service';
+import { TerminalWidgetOptions, TerminalWidget } from '@theia/terminal/lib/browser/terminal-widget';
 import { UriAwareCommandHandler } from '@theia/core/lib/common/uri-command-handler';
 import { FileSystem } from '@theia/filesystem/lib/common';
 import URI from '@theia/core/lib/common/uri';
