+++ packages/xterm-js/src/browser/terminal-preferences.ts
