+++ packages/xterm-js/src/browser/terminal.css
